refactor(keystone/changeset): move OCR3 selector func and add logs
